analyzer_service/requirements.txt
numpy
paho-mqtt